rename types to cast to avoid conflict with lib
