+++ backend/geonature/core/imports/checks/dataframe/__init__.py
@@ -1,4 +1,4 @@
 from .core import *
 from .geometry import check_geometry
-from .types import check_types
+from .cast import check_types
 from .dates import concat_dates
